+++ src/constants.ts
@@ -1,3 +1,14 @@
-export const COMMANDER_SERVICE_TOKEN = 'nests-console.Commander';
+/**
+ * The token to store the root instance of commander
+ */
+export const CLI_SERVICE_TOKEN = 'nests-console.Cli';
+
+/**
+ * The key to store metadata of a Console decorator
+ */
 export const CONSOLE_METADATA_NAME = 'nests-console.Console';
+
+/**
+ * The key to store metadata of a Command decorator
+ */
 export const COMMAND_METADATA_NAME = 'nests-console.Command:';